Bluetooth: MGMT: Fix GET_CLOCK_INFO when reading only local clock
When the Address parameter to GET_CLOCK_INFO is set to BDADDR_ANY, the
command should succeed and return only the local clock value. However,
currently it always fails in this case, because it attempts to read the
piconet clock for invalid address BDADDR_ANY.

This appears to be broken in a refactoring, as the connection used to be
specified in cmd->user_data to get_clock_info_sync, which is no longer
done.

Fix the behavior by doing the comparisons to BDADDR_ANY everywhere
necessary, and not attempting to read piconet clock when not intended.

Fixes: 7b445e220db9 ("Bluetooth: MGMT: Fix holding hci_conn reference while command is queued")
Signed-off-by: Pauli Virtanen <pav@iki.fi>