Bluetooth: keep track of timing and sequence information for ISO
To allow synchronization to proper transmission timing, keep track of
relevant information for ISO connections:

The time of last Number of Completed Packets event, and the sequence
number of the last SDU sent to the controller and the number of packets
queued for that connection, at that point of time.

Also implement LE Read ISO TX Sync command, and keep track of
information provided by its completion event.  Also record then the
previous completed packets information, to allow correlating them.

Signed-off-by: Pauli Virtanen <pav@iki.fi>