Bluetooth: ISO: add ioctl for reading TX sync information
Add an ioctl that allows reading ISO transmission synchronization
information. It informs userspace of:

- The result of a previous LE Read ISO TX Sync command, and the time of
  its completion.

- The time of the Number of Completed Packets event for the connection,
  previous to the sync command completion.

- The values of the sequence number of the last SDU sent to the controller
  and the number of packets queued in the controller for that connection,
  at the time of the number of packets event.

In addition, the ioctl queues a new LE Read ISO TX Sync command to
refresh the values.

The timestamps and sequence numbers in LE Read ISO TX Sync command are
determined by controller, and generally are not in sync with the host,
and may also be bogus (e.g. for Intel AX210 as of March 2023).  The
userspace most likely needs some out-of-band method to access the
controller to make use of these values as absolute timing reference. The
timestamp however provides a way to synchronize different CIS in the
same CIG.

When such direct access to controller is not available, userspace can
use the packet completion event data which is always available.

The other timestamps are in CLOCK_MONOTONIC, which is known to
userspace.

The sequence number allows the userspace to determine which SDUs were
still in the socket buffer. Together with the controller queue size,
this allows determining the total SDU queue size, at the time of the
packet completion event.

With the completion event timing and known SDU interval, the userspace
can then determine the approximate timing when the controller is going
to consider a given SDU handled. This then allows determining when new
data must be supplied, and also synchronization of the packet send
timing so that data is not accidentally submitted to the wrong SDU
interval due to timing jitter.

Signed-off-by: Pauli Virtanen <pav@iki.fi>