Bluetooth: hci_conn: add hci_conn_is_alive
A delayed operation such as hci_sync on a given hci_conn needs to take
hci_conn_get, so that the hci_conn doesn't get freed in the meantime.
This does not guarantee the conn is still alive in a valid state, as it
may be cleaned up in the meantime, so one needs to check if it is still
in conn_hash to know if it's still alive.

Simplify this alive check, we don't need to go through the list to know
if the conn is in the list when we are holding hdev->lock.

Signed-off-by: Pauli Virtanen <pav@iki.fi>