Bluetooth: add support for skb TX timestamping
Track skbs to be TX timestamped in controller queue, and generate TX
timestamps on packet complete event.

The controller queue tracking is a simple ringbuffer holding skb clones,
or NULLs for queued packets not going to be TX timestamped.  It holds
clones, since the driver may concurrently modify skbs sent to it.

Signed-off-by: Pauli Virtanen <pav@iki.fi>